requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
alembic==1.18.4
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asttokens==3.0.1
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.5
category-encoders==2.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cloudpickle==3.1.2
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.7
cuda-bindings==12.9.4
cuda-pathfinder==1.2.2
cuda-toolkit==12.6.3
cycler==0.12.1
databricks-sdk==0.102.0
debugpy==1.8.20
decorator==5.2.1
docker==7.1.0
executing==2.2.1
fastapi==0.135.3
filelock==3.25.2
flask==3.1.3
flask-cors==6.0.2
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
gitdb==4.0.12
gitpython==3.1.46
google-auth==2.49.1
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
greenlet==3.3.2
gunicorn==25.3.0
h11==0.16.0
huey==2.6.0
idna==3.11
importlib-metadata==8.7.1
ipykernel==7.2.0
ipython==9.12.0
ipython-pygments-lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-client==8.8.0
jupyter-core==5.9.1
kiwisolver==1.5.0
mako==1.3.10
markupsafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mlflow==3.11.1
mlflow-skinny==3.11.1
mlflow-tracing==3.11.1
mplcyberpunk==0.7.6
mpmath==1.3.0
multidict==6.7.1
narwhals==2.18.1
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.3
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.85
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.28.9
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.6.77
opentelemetry-api==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
packaging==26.0
pandas==2.3.3
parso==0.8.6
patsy==1.0.2
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.4
prettytable==3.17.0
prompt-toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.6
psutil==7.2.2
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pydeck==0.9.1
pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
pyyaml==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
setuptools==70.2.0
six==1.17.0
skops==0.13.0
smmap==5.0.3
sqlalchemy==2.0.49
sqlparse==0.5.5
stack-data==0.6.3
starlette==1.0.0
statsmodels==0.14.6
streamlit==1.56.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
toml==0.10.2
torch==2.11.0+cu126
torchutils==0.0.4
torchvision==0.26.0+cu126
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
triton==3.6.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2026.1
urllib3==2.6.3
uvicorn==0.44.0
watchdog==6.0.0
wcwidth==0.6.0
werkzeug==3.1.8
yarl==1.23.0
zipp==3.23.0
